Models/requirements.txt
streamlit
fireworks-ai
langchain_fireworks
langchain-anthropic
pillow